backend/requirements.txt
# backend/requirements.txt

absl-py==0.12.0
alembic==1.6.2
astunparse==1.6.3
cachetools==4.2.2
chardet==4.0.0
click==8.0.0
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
flask==2.0.0
flask-cors==3.0.10
flask-login==0.5.0
flask-migrate==3.0.0
flask-sqlalchemy==2.5.1
flask-wtf==0.14.3
flatbuffers==1.12
gast==0.4.0
gunicorn
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
itsdangerous==2.0.0
jinja2==3.0.0
keras==2.2.4
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
mako==1.1.4
markdown==3.3.4
markupsafe==2.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
pdfkit==0.6.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pyyaml==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
sqlalchemy==1.4.15
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.1.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
werkzeug==2.0.0
wrapt==1.12.1
wtforms==2.3.3
fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
